+++ did/did-auth/lib/authenticator/wallet.js
@@ -145,6 +145,8 @@ class WalletAuthenticator extends BaseAuthenticator {
    * @method
    * @param {object} params
    * @param {object} params.claims - info required by application to complete the auth
+   * @param {object} params.pathname - pathname to assemble callback url
+   * @param {object} params.challenge - random challenge to be included in the body
    * @param {object} params.extraParams - extra query params and locale
    * @param {string} params.context.token - action token
    * @param {string} params.context.userDid - decoded from req.query, base58
@@ -154,7 +156,7 @@ class WalletAuthenticator extends BaseAuthenticator {
    * @param {string} params.context.sessionDid - did of logged-in user
    * @returns {object} { appPk, authInfo }
    */
-  async sign({ context, claims, pathname = '', extraParams = {} }) {
+  async sign({ context, claims, pathname = '', challenge = '', extraParams = {} }) {
     const claimsInfo = await this.genRequestedClaims({ claims, context, extraParams });
     const chainInfoParams = Object.assign({}, context, extraParams);
 
@@ -163,6 +165,7 @@ class WalletAuthenticator extends BaseAuthenticator {
 
     const payload = {
       action: 'responseAuth',
+      challenge,
       appInfo: this.appInfo,
       chainInfo: this.getChainInfo(chainInfoParams, tmp ? tmp.chainInfo : undefined),
       requestedClaims: claimsInfo.map(x => {
@@ -218,7 +221,7 @@ class WalletAuthenticator extends BaseAuthenticator {
   verify(data, locale = 'en', enforceTimestamp = true) {
     return new Promise(async (resolve, reject) => {
       try {
-        const { iss, action = 'responseAuth', requestedClaims } = await this._verify(
+        const { iss, challenge = '', action = 'responseAuth', requestedClaims } = await this._verify(
           data,
           'userPk',
           'userInfo',
@@ -232,9 +235,10 @@ class WalletAuthenticator extends BaseAuthenticator {
           userPk: data.userPk,
           claims: requestedClaims,
           action,
+          challenge,
         });
 
-        debug('verify.context', { userPk: data.userPk, userDid: toAddress(iss), action });
+        debug('verify.context', { userPk: data.userPk, userDid: toAddress(iss), action, challenge });
         debug('verify.claims', requestedClaims);
       } catch (err) {
         reject(err);