+++ did/did-auth/tests/handlers/util.spec.js
@@ -1,4 +1,4 @@
-const { parseWalletUA } = require('../../lib/handlers/util');
+const { parseWalletUA, getStepChallenge, preparePathname } = require('../../lib/handlers/util');
 
 describe('#parseWalletUA', () => {
   describe('#android', () => {
@@ -9,9 +9,7 @@ describe('#parseWalletUA', () => {
     });
 
     test('should parse "User-Agent: okhttp/3.12.2 ArcWallet/ (Linux; U; Android 25; LGE Build/Nexus5)" correctly', async () => {
-      const wallet = parseWalletUA(
-        'User-Agent: okhttp/3.12.2 ArcWallet/ (Linux; U; Android 25; LGE Build/Nexus5)'
-      );
+      const wallet = parseWalletUA('User-Agent: okhttp/3.12.2 ArcWallet/ (Linux; U; Android 25; LGE Build/Nexus5)');
 
       expect(wallet).toEqual({ os: 'android', version: '' });
     });
@@ -39,11 +37,57 @@ describe('#parseWalletUA', () => {
     });
 
     test('should parse "ArcWallet/1.3.29 iPhone12,3 iOS/13.0 CFNetwork/1098.7 Darwin/19.0.0" correctly', async () => {
-      const wallet = parseWalletUA(
-        'ArcWallet/1.3.29 iPhone12,3 iOS/13.0 CFNetwork/1098.7 Darwin/19.0.0'
-      );
+      const wallet = parseWalletUA('ArcWallet/1.3.29 iPhone12,3 iOS/13.0 CFNetwork/1098.7 Darwin/19.0.0');
 
       expect(wallet).toEqual({ os: 'ios', version: '1.3.29' });
     });
+  });
+});
+
+describe('#getStepChallenge', () => {
+  test('should return random bytes', async () => {
+    const c1 = getStepChallenge();
+    const c2 = getStepChallenge();
+    expect(c1).toBeTruthy();
+    expect(c2).toBeTruthy();
+    expect(c1).not.toEqual(c2);
+  });
+});
+
+describe('#preparePathname', () => {
+  test('should return correct path when not nested', async () => {
+    const path = '/api/did/login/auth';
+    const req = {
+      originalUrl: '/api/did/login/auth',
+    };
+
+    expect(preparePathname(path, req)).toEqual('/api/did/login/auth');
+  });
+
+  test('should return correct path when not nested: retrieve', async () => {
+    const path = '/api/did/login/retrieve';
+    const req = {
+      originalUrl: '/api/did/login/retrieve',
+    };
+
+    expect(preparePathname(path, req)).toEqual('/api/did/login/retrieve');
+  });
+
+  test('should return correct path when nested', async () => {
+    const path = '/api/did/login/auth';
+    const req = {
+      originalUrl: '/root/nested/api/did/login/auth',
+    };
+
+    expect(preparePathname(path, req)).toEqual('/root/nested/api/did/login/auth');
+  });
+
+  test('should return correct path when netlify', async () => {
+    const path = '/api/did/login/auth';
+    const req = {
+      originalUrl: '/.netlify/functions/app/api/did/login/auth',
+    };
+
+    expect(preparePathname(path, req)).toEqual('/.netlify/functions/app/api/did/login/auth');
   });
 });